fix: compare bounds in i64 space, not after casting, for 32-bit targets
Review finding on PR #58: every bound check added across the 13 files
used the shape `(*end as usize) > right_len` -- casting the i64 value
down to usize *before* comparing. On a 32-bit target (this crate's
release matrix includes x86 and armv7, per .github/workflows/release.yml)
usize is 32 bits, so a genuinely oversized end (e.g. 2**32 + 1) truncates
to a small value before the check ever sees it.

In most of the 11 non-posns files this produced silently wrong/truncated
results rather than a panic, since the same truncated value was used
consistently for both the check and the actual loop bound (self-
consistent but wrong). Two places were worse:

- comp_first_starts.rs / comp_ne_starts_1st.rs had no explicit upper-bound
  check at all -- they relied on "a Range with start > end is naturally
  empty" as the safety net, which breaks under wraparound: a huge start
  can truncate to something *smaller* than end, turning an out-of-range
  row into a seemingly valid one.
- comp_posns.rs / comp_posns_ne.rs's indexer check could let a large
  positive indexer wrap to an in-range value and silently select the
  wrong row from `right` -- the reviewer's specific example.

Fix: compare in i64 space (`right_len as i64`) instead of casting the i64
value to usize first -- a lossless widening cast on the length side,
rather than a lossy narrowing cast on the value side. Same cost as before
(one comparison either way), so this doesn't reopen #55's checked_range
perf finding. comp_first_starts.rs/comp_ne_starts_1st.rs additionally gained
the explicit upper-bound check they were missing.

Not independently regression-tested: this class of bug only manifests
when usize is narrower than i64, and this repo's CI only runs `cargo
test` on ubuntu-latest (x86_64) -- the 32-bit release targets are
cross-compiled but never test-executed. Verified instead by re-running
the full suite (212 passed) and a rebuilt wheel end-to-end on this (64-bit)
machine to confirm no regression in ordinary behavior.

+++ src/compare/comp_ends.rs
@@ -36,9 +36,15 @@ macro_rules! generic_compare {
             // way walking `right_array[nn]` out of bounds once the main
             // loop reaches it.
             let right_len = right_array.len();
+            // Comparing in i64 space (`right_len as i64`, a lossless
+            // widening cast) rather than casting `end` down to `usize`
+            // first: on a 32-bit target `usize` is 32 bits, so a
+            // genuinely oversized `end` (e.g. 2^32 + 1) would truncate to
+            // a small value *before* this check ever saw it, silently
+            // passing validation instead of being rejected.
             if let Some(bad_end) = ends_array
                 .iter()
-                .find(|e| **e < 0 || (**e as usize) > right_len)
+                .find(|e| **e < 0 || **e > right_len as i64)
             {
                 return Err(PyValueError::new_err(format!(
                     "end must be within 0..={right_len}; got {bad_end}"
